backend/requirements.txt
# SecureVault Backend Dependencies - Production Ready

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Storage (S3-compatible)
minio==7.2.0

# Search & Logging
elasticsearch==8.11.0

# Security & Authentication
PyJWT==2.8.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2

# Environment
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Security Scanning
bandit==1.7.5
safety==3.2.0